+++ inferno-vc/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Revision History for inferno-vc
 
+## 0.1.2 -- 2023-02-09
+* Remove Arbitrary and ToADTArbitrary instances
+
 ## 0.1.0.0 -- 2022-11-28
 * Prepare for OSS release